requirements.txt
alembic==1.12.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fhir.resources==7.0.2
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pydantic==1.10.13
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3.post1
requests==2.31.0
sendgrid==6.10.0
six==1.16.0
SQLAlchemy==2.0.23
starkbank-ecdsa==2.2.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
WTForms==3.1.1
